+++ Plugin/Uplink/Source/UplinkEditor/Private/UplinkTools.h
@@ -35,5 +35,5 @@ namespace UplinkTools
 	void RegisterAutoplay(FUplinkToolRegistry& Registry);
 	void RegisterRepair(FUplinkToolRegistry& Registry);
 	void RegisterReferences(FUplinkToolRegistry& Registry);
-	void RegisterSemantic(FUplinkToolRegistry& Registry);
+	void RegisterSemantic(FUplinkToolRegistry& Registry, FUplinkEventRecorder& Recorder);
 }